[BackEnd] refactor : 웹소켓 에러 디렉토리 생성 및 핸들러,dto들 이동
